Refactor Span interfaces (#216)
Define immutable schema of a Span in `SpanData` and have `ReadableSpan` extend from it. The former will have the guarantee of immutability, while the latter takes that away.

In turn, both will return Span Events and Links that are immutable, which used to be called ReadableEvent and ReadableLink before, but has been renamed *Data for consistency.

As well, the status of a Span was changed to encompass both the status code and description. Currently, the description for when the status is `OK` or `UNSET` are ignored, so I'm preserving this behaviour for now - I will refactor that in a subsequent PR.