Consolidate interfaces of links, span events, and attributes (#258)
The read-only getters for the current set of events, links, and attributes are defined in a bunch of different places. 

Consolidate this by putting the event and link getters in `SpanSchema` and the setters on attributes in a new interface called `MutableAttributeContainer` which extends from the slimmed down `AttributeContainer`

Doing this allows for the `Link` and `SpanEvent` interfaces to extend from `LinkData` and `EventData`, further reducing duplication.