Add tests for SpanProcessor adapter (#278)
Add to the `Span` interface that notion of the span's context at creation time. This is so it could be used in the API to create a span with this given span as a parent.

We can probably use a the root context as a basis but this is probably more complete.